requirements.txt
Django==1.5.4
Markdown==2.3.1
MySQL-python==1.2.3
South==0.8.2
Unipath==1.0
boto==2.13.3
django-bootstrap3==1.0.0
django-pagedown==0.0.4
django-storages==1.1.8
django-taggit==0.10
django-voting==0.2
wsgiref==0.1.2
